requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.0.1
certifi==2025.6.15
click==8.2.1
colorama==0.4.6
deprecation==2.1.0
dnspython==2.7.0
dotenv==0.9.9
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.115.14
gotrue==2.12.2
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
lxml==6.0.0
packaging==25.0
passlib==1.7.4
pluggy==1.6.0
postgrest==1.1.1
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
PyPDF2==3.0.1
pytest==8.4.1
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.20
realtime==2.5.3
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
storage3==0.12.0
StrEnum==0.4.15
supabase==2.16.0
supafunc==0.10.1
tomli==2.2.1
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.35.0
websockets==15.0.1
